Separate all buffers in modules for an easier work with code
